requirements.txt
Django>=5.2,<6.0
Pillow>=10.0
stripe>=11.0
